# Fix file descriptors for v1 StorageFixtures [Partially fixes the problem]