--- a/build_tooling/monitor_opened_files.sh
+++ b/build_tooling/monitor_opened_files.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+LOGFILE="opened_files.log"
+PROCESS_NAME="pytest"
+
+# Clear or create the log file
+: > "$LOGFILE"
+
+while true; do
+    timestamp=$(date +"%Y-%m-%d %H:%M:%S")
+    count=$(lsof -c "$PROCESS_NAME" 2>/dev/null | wc -l)
+    echo "$timestamp $count" >> "$LOGFILE"
+    sleep 10
+done
+