--- a/build_tooling/monitor_opened_files.sh
+++ b/build_tooling/monitor_opened_files.sh
@@ -18,6 +18,10 @@ while true; do
 
     for pid in $(pgrep -f 'pytest|multiprocessing'); do
         count=$(lsof -p "$pid" 2>/dev/null | wc -l)
+        # try with sudo if count is 0
+        if [ "$count" -eq 0 ]; then
+            count=$(sudo lsof -p "$pid" 2>/dev/null | wc -l)
+        fi
         output+=" $count"
     done