--- a/build_tooling/monitor_opened_files.sh
+++ b/build_tooling/monitor_opened_files.sh
@@ -1,13 +1,11 @@
 #!/bin/bash
 
+# This file will contain number of opened files for each process we need 
+# to track over time
 LOGFILE="${1:-opened_files.log}"
-PROCESSESFILE="${2:-active_processes.log}"
-
-PROCESS_NAME="pytest"
 
 # Clear or create the log file
 : > "$LOGFILE"
-: > "$PROCESSESFILE"
 
 while true; do
 
@@ -17,24 +15,16 @@ while true; do
     output="$timestamp"
 
     for pid in $(pgrep -f 'pytest|multiprocessing'); do
-        echo "PID detected: $pid"
-        ps -p $pid -o pid,ppid,cmd,%mem,%cpu,etime
-        count=$(sudo ls  /proc/$pid/fd 2>/dev/null | wc -l)
+        count=$(ls  /proc/$pid/fd 2>/dev/null | wc -l)
         # try with sudo if count is 0
         if [ "$count" -eq 0 ]; then
             echo "Use fallback method"
-            count=$(sudo lsof -p "$pid" 2>/dev/null | wc -l)
+            count=$(lsof -p "$pid" 2>/dev/null | wc -l)
         fi        
         output+=" $count"
     done
 
     echo -e "$output" >> "$LOGFILE"
 
-    # Processes dump to file
-    echo $timestamp >> "$PROCESSESFILE"
-    ps -ef >> "$PROCESSESFILE"
-    echo "======================================" >> "$PROCESSESFILE"
-
-    sleep 10
 done