--- a/build_tooling/parallel_test.sh
+++ b/build_tooling/parallel_test.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 set -e
 
+# Save firectory from which we started the script to go back after it finishes
+pushd "$(pwd)" >/dev/null
+
 tooling_dir="$(dirname $BASH_SOURCE)"
 echo Saving results to ${TEST_OUTPUT_DIR:="$(realpath "$tooling_dir/../cpp/out")"}
 [[ -d "$TEST_OUTPUT_DIR" ]] || mkdir -p "$TEST_OUTPUT_DIR"
@@ -35,3 +38,5 @@ else
         --basetemp="$PARALLEL_TEST_ROOT/temp-pytest-output" \
         $PYTEST_ADD_TO_COMMAND_LINE $ARCTICDB_PYTEST_ARGS 2>&1
 fi
+
+popd >/dev/null