--- a/python/arcticdb/storage_fixtures/api.py
+++ b/python/arcticdb/storage_fixtures/api.py
@@ -97,7 +97,10 @@ class StorageFixture(_SaferContextManager):
         suffix = 0
         while f"{name}{suffix or ''}" in libs_from_factory:
             suffix += 1
-        libs_from_factory[f"{name}{suffix or ''}"] = out
+        # No caching for NativeVersionStores at StorageFixture
+        # Preserve only names of created NativeVersionStores
+        # libs_from_factory[f"{name}{suffix or ''}"] = out
+        libs_from_factory[f"{name}{suffix or ''}"] = None
         return out
 
     def create_version_store_factory(self, default_lib_name: str, **defaults) -> Callable[..., NativeVersionStore]:
@@ -125,8 +128,9 @@ class StorageFixture(_SaferContextManager):
 
     def slow_cleanup(self, failure_consequence=""):
         for lib in self.libs_from_factory.values():
-            with handle_cleanup_exception(self, lib, consequence=failure_consequence):
-                lib.version_store.clear()
+            if lib is not None:
+                with handle_cleanup_exception(self, lib, consequence=failure_consequence):
+                    lib.version_store.clear()
         self.libs_from_factory.clear()
 
         arctic = self.create_arctic()