--- a/python/arcticdb/storage_fixtures/s3.py
+++ b/python/arcticdb/storage_fixtures/s3.py
@@ -21,7 +21,7 @@ import random
 from datetime import datetime
 
 import requests
-from typing import Optional, Any, Type
+from typing import Optional, Any, Type, Union
 
 import werkzeug
 import botocore.exceptions
@@ -38,6 +38,7 @@ from .utils import (
 from arcticc.pb2.storage_pb2 import EnvironmentConfigsMap
 from arcticdb.version_store.helper import add_gcp_library_to_env, add_s3_library_to_env
 from arcticdb_ext.storage import AWSAuthMethod, NativeVariantStorage, GCPXMLSettings as NativeGCPXMLSettings
+from arcticdb_ext.tools import S3Tool
 
 # All storage client libraries to be imported on-demand to speed up start-up of ad-hoc test runs
 
@@ -252,6 +253,19 @@ class GcpS3Bucket(S3Bucket):
         return cfg, self.native_config
 
 
+def check_bucket(sff: Union['BaseS3StorageFixtureFactory', 'BaseGCPStorageFixtureFactory']):
+    s3_tool = S3Tool(sff.default_bucket, sff.default_key.user_name, sff.default_key.secret, sff.endpoint)
+    content = s3_tool.list_bucket(sff.default_prefix)
+
+    logger.warning(f"Total objects left: {len(content)}")
+    logger.warning(f"First 100: {content[0:100]}")
+    left_from = set()
+    for key in content:
+        library_name = key.split("/")[1] # get the name from object
+        left_from.add(library_name)
+    logger.warning(f"Left overs from libraries: {left_from}")
+    assert len(content) < 1
+
 class BaseS3StorageFixtureFactory(StorageFixtureFactory):
     """Logic and fields common to real and mock S3"""
 
@@ -302,6 +316,8 @@ class BaseS3StorageFixtureFactory(StorageFixtureFactory):
             # and if we try to delete the bucket, it will fail
             b.slow_cleanup(failure_consequence="The following delete bucket call will also fail. ")
 
+            check_bucket(self)
+
 
 class BaseGCPStorageFixtureFactory(StorageFixtureFactory):
     """Logic and fields common to real and mock S3"""
@@ -335,6 +351,8 @@ class BaseGCPStorageFixtureFactory(StorageFixtureFactory):
             # We are not writing to buckets in this case
             # and if we try to delete the bucket, it will fail
             b.slow_cleanup(failure_consequence="The following delete bucket call will also fail. ")
+
+            check_bucket(self)
 
 
 def real_s3_from_environment_variables(