--- a/python/arcticdb/util/utils.py
+++ b/python/arcticdb/util/utils.py
@@ -115,14 +115,14 @@ def delete_nvs(nvs: NativeVersionStore, ac: Arctic = None):
 
 def delete_library(ac: Arctic, lib_name: str):
     try:
-        ac.create_library(lib_name)
+        ac.delete_library(lib_name)
     except arcticdb_ext.exceptions.StorageException as e:
         if "LMDB" in repr(e): # skip LMDB errors
             pass
         else: 
             logger.warning(f"Caught exception during deletion of library '{lib_name}' - {repr(e)}")
             raise
-            
+
 
 class  TimestampNumber:
     """