#### Reference Issues/PRs
<!--Example: Fixes #1234. See also #3456.-->

#### What does this implement or fix?

(IMPORTANT:  it is important to note that this approach reduces the file descriptors usade with 40-50% it is not sure it eliminates all leakages, just does that from the test libraries caches, other leakages might still be present)

The main goal of this PR is to limit the number of file descriptors that are needed in order to run tests on Linux and MacOs as those OS-es have file descriptor limits. Windows does not have file descriptor limits.

The issue we have is with StorageFixture class in api.py - it caches all NativeVersionStore objects it creates when ```create_version_store_factory``` method is invoked. As our storage factories run now in session scope, the lifespan of this cache is the duration of tests. Thus for the duration of tests we have objects that keep their connections to storages. Additionally the storage space is nor wiped after the the test execution but only at the end. Thus if there is catatstrophic failure preventing reaching the end the storage will not be freed.

Approch taken is to shorten lifespan of the NativeVersionStores created for each test only for the duration of the test. Wherever it is possible the changes were implemented in the fixtures -  each fixture that creates NVS object is in charge of cleaning it now. For factory fixtures that is not possible and therefore tests were modified to have ```try ... finally``` clause.

There are also 2 helper utilities ```delete_nvs``` and ```delete_library``` implemented which help taking care of some exceptions that may occur during that process that would otherwise trigger failed or error state during cleanup

Also during the development it was noticed that LMDB tests and other types of tests rarely cleanup after themselves. That manifested in evergrowing disk space used by tests that leads to out of space errors for runners. Therefore this PR also tries to address that problem that exists with such tests and is not limited to v1 API fixtures only but also with v2 API fixtures

This PR also introduces a background process that will collect info fo file descriptor numbers during the execution of tests and at the end it will print the maximim number of file descriptors that we reached. This monitoring is needed to objectivly measure and understand needs of tests during test execution. In future this approach could be used to track much more stats about test execution. (NOTE: works only on local machines for now. Does not work on github unfortunately)

* Test Results (locally on WSL)

EXECUTION WITH FIX GCP/AWS_S3 TESTS FROM: python/tests/unit/arcticdb/version_store/test_parallel.py

All Tests finished after fix on a machine with limit 1024 file descriptors
 
During execution opened files reached 893 for GCP (for AWS S3 tests it was 725, or approx 23% more opened files on same tests for GCP bs AWS S3)

As the tests required arround 1550 opened files  for GCP when executed without this fix we have more than 45% reduction of file descriptors usage

NOTE: For monitoring following script was used:
``` 
while true; do
  lsof -c pytest | wc -l
  sleep 10
done
``` 

IMPORTANT: Because with the fix each NativeVersionStore was deleted immediately after usage all storage space was freed as soon as possible.  That is contrary to what was used to be before when storage was deleted at the end. And if file descriptors are exhausted before that no storage can be reclaimed

** Performance Analysis of code changes

Here we compare the execution of one and the same test suite with old and with new code in GitHub

- 3.8 Linux  - integration tests
-- old code - https://github.com/man-group/ArcticDB/actions/runs/15825356207/job/44604755311
```
= 3139 passed, 1356 skipped, 90 xfailed, 4 xpassed, 14 warnings in 1481.28s (0:24:41) =
```

-- NEW CODE -https://github.com/man-group/ArcticDB/actions/runs/15824842292/job/44602814883
```
= 3121 passed, 1357 skipped, 90 xfailed, 4 xpassed, 10 warnings in 1413.03s (0:23:33) =
```

- 3.11 Windows - integration tests

-- old code https://github.com/man-group/ArcticDB/actions/runs/15825356207/job/44605215726
```
= 2806 passed, 1655 skipped, 86 xfailed, 4 xpassed, 3740 warnings in 3413.78s (0:56:53) =
```

-- NEW CODE: https://github.com/man-group/ArcticDB/actions/runs/15824842292/job/44603214947
```
= 2788 passed, 1656 skipped, 86 xfailed, 4 xpassed, 3731 warnings in 3263.39s (0:54:23) =
```

- 3.13 Linux - integration tests
- old code - https://github.com/man-group/ArcticDB/actions/runs/15825356207/job/44604764324
```
= 3149 passed, 1346 skipped, 90 xfailed, 4 xpassed, 7647 warnings in 1337.87s (0:22:17) =
```

--NEW CODE - https://github.com/man-group/ArcticDB/actions/runs/15824842292/job/44602827969
```
= 3131 passed, 1347 skipped, 90 xfailed, 4 xpassed, 7642 warnings in 1374.96s (0:22:54) =
```

Analysis: there are no evidences that new fixtures are degrading the performance. The code at par with previous one and sligly faster which can be due to lower (18) number of tests or the code changes itself. 

*** Unit tests performance analysis

The only notable degradation is observed at unit tests level:
- In the past unit tests used to run for 5 minutes https://github.com/man-group/ArcticDB/actions/runs/15825356207/job/44604768352
- Currently they are running for 15 minutes: https://github.com/man-group/ArcticDB/actions/runs/15824842292/job/44602861927

The main difference is adding reclaiming of space for unit tests. The old code did not had it as majority tests were LMDB tests


#### Any other comments?

#### Checklist

<details>
  <summary>
   Checklist for code changes...
  </summary>
 
 - [ ] Have you updated the relevant docstrings, documentation and copyright notice?
 - [ ] Is this contribution tested against [all ArcticDB's features](../docs/mkdocs/docs/technical/contributing.md)?
 - [ ] Do all exceptions introduced raise appropriate [error messages](https://docs.arcticdb.io/error_messages/)?
 - [ ] Are API changes highlighted in the PR description?
 - [ ] Is the PR labelled as enhancement or bug so it appears in autogenerated release notes?
</details>

<!--
Thanks for contributing a Pull Request to ArcticDB! Please ensure you have taken a look at:
 - ArcticDB's Code of Conduct: https://github.com/man-group/ArcticDB/blob/master/CODE_OF_CONDUCT.md
 - ArcticDB's Contribution Licensing: https://github.com/man-group/ArcticDB/blob/master/docs/mkdocs/docs/technical/contributing.md#contribution-licensing
-->