@phoebusm 2025-06-26
> As the tests required arround 1550 opened files for GCP when executed without this fix we have more than 45% reduction of file descriptors usage

So I guess it's only a partial fix?