@maxim-morozov 2025-06-30
The new PR is in https://github.com/man-group/ArcticDB/pull/2431